api/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=api/requirements/requirements.txt api/requirements/requirements.in
#
aiohttp==3.8.5
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   langchain
    #   starlette
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   pip-tools
    #   uvicorn
dataclasses-json==0.6.1
    # via langchain
dnspython==2.4.2
    # via pymongo
fastapi==0.103.2
    # via -r api/requirements/requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.0
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.306
    # via -r api/requirements/requirements.in
langsmith==0.0.41
    # via langchain
loguru==0.7.2
    # via -r api/requirements/requirements.in
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.7
    # via langchain
numpy==1.26.0
    # via
    #   langchain
    #   numexpr
openai==0.28.1
    # via -r api/requirements/requirements.in
packaging==23.2
    # via
    #   build
    #   marshmallow
pip-tools==7.3.0
    # via -r api/requirements/requirements.in
pydantic==2.4.2
    # via
    #   fastapi
    #   langchain
    #   langsmith
pydantic-core==2.10.1
    # via pydantic
pymongo==4.5.0
    # via -r api/requirements/requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r api/requirements/requirements.in
pyyaml==6.0.1
    # via langchain
requests==2.31.0
    # via
    #   langchain
    #   langsmith
    #   openai
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.21
    # via langchain
starlette==0.27.0
    # via fastapi
tenacity==8.2.3
    # via langchain
tqdm==4.66.1
    # via openai
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.6
    # via requests
uvicorn==0.23.2
    # via -r api/requirements/requirements.in
wheel==0.41.2
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
